# device-farm/infra/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Device Farm API
  description: 设备农场管理平台 API 文档
  version: 1.0.0
  contact:
    name: Device Farm Team
    email: device-farm@example.com
servers:
  - url: http://localhost:80
    description: 开发环境
  - url: https://device-farm.example.com
    description: 生产环境
...
paths:
...
  /api/v1/devices:
    get:
      tags:
        - devices
      summary: 获取设备列表
      description: 获取所有设备列表，支持按状态、平台筛选
      operationId: listDevices
      parameters:
        - name: status
          in: query
          description: 设备状态
          schema:
            type: string
            enum: [online, offline, busy, maintenance]
        - name: platform
          in: query
          description: 设备平台
          schema:
            type: string
            enum: [android, ios]
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: 每页数量
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceListResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - devices
      summary: 注册新设备
      description: 注册一个新的测试设备
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegisterRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: 设备序列号已存在
  /api/v1/devices/{deviceId}:
    get:
      tags:
        - devices
      summary: 获取设备详情
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - devices
      summary: 更新设备信息
      operationId: updateDevice
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - devices
      summary: 删除设备
      operationId: deleteDevice
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '204':
          description: 删除成功
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/devices/{deviceId}/release:
    post:
      tags:
        - devices
      summary: 释放设备
      description: 释放设备，设备状态恢复为online
      operationId: releaseDevice
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 释放原因
      responses:
        '200':
          description: 释放成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== 投屏控制 ====================
...
  /api/v1/reports:
    get:
      tags:
        - reports
      summary: 获取报告列表
      operationId: listReports
      parameters:
        - name: taskId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
  /api/v1/reports/{reportId}:
    get:
      tags:
        - reports
      summary: 获取报告详情
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/reportId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

# ==================== 组件定义 ====================
components:
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: 设备ID
      schema:
        type: string
...
    reportId:
      name: reportId
      in: path
      required: true
      description: 报告ID
      schema:
        type: string
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
        serial:
          type: string
        platform:
          type: string
          enum: [android, ios]
        model:
          type: string
        brand:
          type: string
        osVersion:
          type: string
        screenWidth:
          type: integer
        screenHeight:
          type: integer
        status:
          type: string
          enum: [online, offline, busy, maintenance]
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceRegisterRequest:
      type: object
      required:
        - serial
        - platform
      properties:
        serial:
          type: string
        platform:
          type: string
          enum: [android, ios]
        model:
          type: string
        brand:
          type: string
        osVersion:
          type: string
        screenWidth:
          type: integer
        screenHeight:
          type: integer
    DeviceUpdateRequest:
      type: object
      properties:
        model:
          type: string
        brand:
          type: string
        osVersion:
          type: string
        status:
          type: string
          enum: [online, offline, busy, maintenance]
    DeviceResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/Device'
    DeviceListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/Device'

    # 投屏相关
...
    Report:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        totalCases:
          type: integer
        passedCases:
          type: integer
        failedCases:
          type: integer
        skippedCases:
          type: integer
        duration:
          type: integer
          description: 执行时长(秒)
        videoUrl:
          type: string
        logUrl:
          type: string
        reportUrl:
          type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    CaseResult:
      type: object
      properties:
        id:
          type: string
        reportId:
          type: string
        caseName:
          type: string
        status:
          type: string
          enum: [passed, failed, skipped]
        duration:
          type: integer
          description: 执行时长(毫秒)
        errorMessage:
          type: string
        stackTrace:
          type: string
        createdAt:
          type: string
          format: date-time
    ReportDetailResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            report:
              $ref: '#/components/schemas/Report'
            caseResults:
              type: array
              items:
                $ref: '#/components/schemas/CaseResult'
    ReportListResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            items:
              type: array
              items:
                $ref: '#/components/schemas/Report'

    # 错误响应
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'